requirements.txt
# Fleetclaw Configuration Generator Dependencies
# Install with: pip install -r requirements.txt

pyyaml>=6.0
jinja2>=3.0
pydantic>=2.0
ruamel.yaml>=0.17
